streampulse-backup-20251229/requirements.txt
# 수집 & 데이터
requests
schedule
duckdb
psycopg2-binary
pandas

# 에이전트 (LangGraph)
langchain
langgraph
langchain-openai
langchain-community
duckduckgo-search

# 웹/대시보드
streamlit
fastapi
uvicorn
python-dotenv

plotly
sqlalchemy
alembic